refactor(db): split sqlite-agent-integration-repository into sub-modules
Decompose the 952-line god node using composition pattern while preserving
public import paths. Closes #610.
