+++ packages/memento-core/src/infrastructure/database/repositories/sqlite-agent-integration-repository.ts
@@ -1,5 +1,4 @@
 import type Database from 'better-sqlite3';
-import { randomUUID } from 'crypto';
 import type {
   AgentIntegrationRepository,
   CreateAgentMemoryPromotionCandidateInput,
@@ -10,211 +9,31 @@ import type {
   AgentMemoryPromotionCandidate,
   AgentObservation,
   AgentSession,
-  AgentSessionObservationAggregate,
-  AgentSessionPage,
   MemoryProvenance,
   ObservationPage,
   PersistedAgentEventInput,
 } from '../../../domains/agent-integration/types.js';
-
-type SessionRow = {
-  id: string;
-  adapter_name: string;
-  adapter_version: string;
-  contract_version: number;
-  owner_id: string | null;
-  project_id: string | null;
-  process_id: string | null;
-  status: AgentSession['status'];
-  started_at: string;
-  ended_at: string | null;
-  last_event_at: string;
-  max_sequence_no: number;
-  agent_metadata_json: string | null;
-  summary_memory_id: string | null;
-  degraded_reason: string | null;
-  created_at: string;
-  updated_at: string;
-};
-
-type ObservationRow = {
-  id: string;
-  adapter_name: string;
-  event_id: string;
-  session_id: string;
-  event_type: AgentObservation['eventType'];
-  sequence_no: number;
-  tool_name: string | null;
-  outcome: string | null;
-  payload_json: string | null;
-  payload_sha256: string;
-  redaction_metadata_json: string;
-  status: AgentObservation['status'];
-  drop_reason: string | null;
-  late_arrival: number;
-  occurred_at: string;
-  received_at: string;
-  expires_at: string | null;
-};
-
-type ProvenanceRow = {
-  id: string;
-  memory_id: string;
-  session_id: string | null;
-  observation_id: string | null;
-  derivation_type: string;
-  source_deleted: number;
-  created_at: string;
-};
-
-type PromotionCandidateRow = {
-  id: string;
-  fingerprint: string;
-  session_id: string;
-  summary_memory_id: string;
-  target_type: AgentMemoryPromotionCandidate['targetType'];
-  category: AgentMemoryPromotionCandidate['category'];
-  content: string;
-  confidence: number;
-  evidence_observation_ids_json: string;
-  merge_target_memory_id: string | null;
-  status: AgentMemoryPromotionCandidate['status'];
-  memory_id: string | null;
-  rejection_reason: string | null;
-  created_at: string;
-  updated_at: string;
-  reviewed_at: string | null;
-};
-
-function mapSession(row: SessionRow): AgentSession {
-  return {
-    id: row.id,
-    adapterName: row.adapter_name,
-    adapterVersion: row.adapter_version,
-    contractVersion: row.contract_version,
-    ownerId: row.owner_id,
-    projectId: row.project_id,
-    processId: row.process_id,
-    status: row.status,
-    startedAt: row.started_at,
-    endedAt: row.ended_at,
-    lastEventAt: row.last_event_at,
-    maxSequenceNo: row.max_sequence_no,
-    agentMetadataJson: row.agent_metadata_json,
-    summaryMemoryId: row.summary_memory_id,
-    degradedReason: row.degraded_reason,
-    createdAt: row.created_at,
-    updatedAt: row.updated_at,
-  };
-}
-
-function mapObservation(row: ObservationRow): AgentObservation {
-  return {
-    id: row.id,
-    adapterName: row.adapter_name,
-    eventId: row.event_id,
-    sessionId: row.session_id,
-    eventType: row.event_type,
-    sequenceNo: row.sequence_no,
-    toolName: row.tool_name,
-    outcome: row.outcome,
-    payloadJson: row.payload_json,
-    payloadSha256: row.payload_sha256,
-    redactionMetadataJson: row.redaction_metadata_json,
-    status: row.status,
-    dropReason: row.drop_reason,
-    lateArrival: row.late_arrival === 1,
-    occurredAt: row.occurred_at,
-    receivedAt: row.received_at,
-    expiresAt: row.expires_at,
-  };
-}
-
-function mapProvenance(row: ProvenanceRow): MemoryProvenance {
-  return {
-    id: row.id,
-    memoryId: row.memory_id,
-    sessionId: row.session_id,
-    observationId: row.observation_id,
-    derivationType: row.derivation_type,
-    sourceDeleted: row.source_deleted === 1,
-    createdAt: row.created_at,
-  };
-}
-
-function mapPromotionCandidate(row: PromotionCandidateRow): AgentMemoryPromotionCandidate {
-  return {
-    id: row.id,
-    fingerprint: row.fingerprint,
-    sessionId: row.session_id,
-    summaryMemoryId: row.summary_memory_id,
-    targetType: row.target_type,
-    category: row.category,
-    content: row.content,
-    confidence: row.confidence,
-    evidenceObservationIds: JSON.parse(row.evidence_observation_ids_json) as string[],
-    mergeTargetMemoryId: row.merge_target_memory_id,
-    status: row.status,
-    memoryId: row.memory_id,
-    rejectionReason: row.rejection_reason,
-    createdAt: row.created_at,
-    updatedAt: row.updated_at,
-    reviewedAt: row.reviewed_at,
-  };
-}
-
-function normalizePromotionContent(value: string): string {
-  return value.trim().replace(/\s+/g, ' ').toLowerCase();
-}
-
-function encodeCursor(item: AgentObservation): string {
-  return Buffer.from(JSON.stringify([
-    item.sequenceNo,
-    item.occurredAt,
-    item.receivedAt,
-    item.id,
-  ])).toString('base64url');
-}
-
-function decodeCursor(cursor: string): [number, string, string, string] {
-  const parsed = JSON.parse(Buffer.from(cursor, 'base64url').toString('utf8')) as unknown;
-  if (!Array.isArray(parsed) || parsed.length !== 4) {
-    throw new Error('Invalid observation cursor');
-  }
-  return parsed as [number, string, string, string];
-}
-
-function encodeSessionCursor(session: AgentSession): string {
-  return Buffer.from(JSON.stringify([session.lastEventAt, session.id])).toString('base64url');
-}
-
-function decodeSessionCursor(cursor: string): [string, string] {
-  const parsed = JSON.parse(Buffer.from(cursor, 'base64url').toString('utf8')) as unknown;
-  if (
-    !Array.isArray(parsed)
-    || parsed.length !== 2
-    || typeof parsed[0] !== 'string'
-    || typeof parsed[1] !== 'string'
-  ) {
-    throw new Error('Invalid session cursor');
-  }
-  return [parsed[0], parsed[1]];
-}
-
-function emptySessionObservationAggregate(): AgentSessionObservationAggregate {
-  return {
-    total: 0,
-    late: 0,
-    byEventType: {},
-    byStatus: {},
-    redacted: 0,
-    dropped: 0,
-    degraded: 0,
-  };
-}
+import { AgentIntegrationObservationStore } from './agent-integration-observation-store.js';
+import { AgentIntegrationPromotionStore } from './agent-integration-promotion-store.js';
+import { AgentIntegrationProvenanceStore } from './agent-integration-provenance-store.js';
+import { AgentIntegrationSessionStore } from './agent-integration-session-store.js';
 
 export class SqliteAgentIntegrationRepository implements AgentIntegrationRepository {
-  constructor(private readonly db: Database.Database) {}
+  private readonly sessions: AgentIntegrationSessionStore;
+  private readonly observations: AgentIntegrationObservationStore;
+  private readonly provenance: AgentIntegrationProvenanceStore;
+  private readonly promotions: AgentIntegrationPromotionStore;
+
+  constructor(private readonly db: Database.Database) {
+    this.sessions = new AgentIntegrationSessionStore(db);
+    this.observations = new AgentIntegrationObservationStore(db);
+    this.provenance = new AgentIntegrationProvenanceStore(db);
+    this.promotions = new AgentIntegrationPromotionStore(
+      db,
+      id => this.sessions.get(id),
+      operation => this.runInTransaction(operation),
+    );
+  }
 
   runInTransaction<T>(operation: () => T): T {
     return this.db.transaction(operation)();
@@ -232,36 +51,11 @@ export class SqliteAgentIntegrationRepository implements AgentIntegrationReposit
   }
 
   createSession(event: PersistedAgentEventInput, now: string): AgentSession {
-    this.db.prepare(`
-      INSERT INTO agent_session (
-        id, adapter_name, adapter_version, contract_version,
-        owner_id, project_id, process_id, status,
-        started_at, last_event_at, max_sequence_no,
-        agent_metadata_json, created_at, updated_at
-      ) VALUES (?, ?, ?, ?, ?, ?, ?, 'ACTIVE', ?, ?, ?, ?, ?, ?)
-    `).run(
-      event.sessionId,
-      event.adapterName,
-      event.adapterVersion,
-      event.contractVersion,
-      event.scope.ownerId ?? null,
-      event.scope.projectId ?? null,
-      event.scope.processId ?? null,
-      event.occurredAt,
-      event.occurredAt,
-      event.sequenceNo,
-      event.payloadJson,
-      now,
-      now,
-    );
-    return this.getSession(event.sessionId)!;
+    return this.sessions.create(event, now);
   }
 
   getSession(id: string): AgentSession | null {
-    const row = this.db.prepare('SELECT * FROM agent_session WHERE id = ?').get(id) as
-      | SessionRow
-      | undefined;
-    return row ? mapSession(row) : null;
+    return this.sessions.get(id);
   }
 
   listSessions(
@@ -273,408 +67,74 @@ export class SqliteAgentIntegrationRepository implements AgentIntegrationReposit
       ownerId?: string;
       projectId?: string;
     } = {},
-  ): AgentSessionPage {
-    const limit = Math.min(Math.max(query.limit ?? 50, 1), 100);
-    const conditions: string[] = [];
-    const parameters: unknown[] = [];
-    if (query.status) {
-      conditions.push('status = ?');
-      parameters.push(query.status);
-    }
-    if (query.adapterName) {
-      conditions.push('adapter_name = ?');
-      parameters.push(query.adapterName);
-    }
-    if (query.ownerId) {
-      conditions.push('owner_id = ?');
-      parameters.push(query.ownerId);
-    }
-    if (query.projectId) {
-      conditions.push('project_id = ?');
-      parameters.push(query.projectId);
-    }
-    if (query.cursor) {
-      const [lastEventAt, id] = decodeSessionCursor(query.cursor);
-      conditions.push('(last_event_at < ? OR (last_event_at = ? AND id < ?))');
-      parameters.push(lastEventAt, lastEventAt, id);
-    }
-
-    const where = conditions.length > 0
-      ? `WHERE ${conditions.map(condition => `(${condition})`).join(' AND ')}`
-      : '';
-    const rows = this.db
-      .prepare(`
-        SELECT * FROM agent_session
-        ${where}
-        ORDER BY last_event_at DESC, id DESC
-        LIMIT ?
-      `)
-      .all(...parameters, limit + 1) as SessionRow[];
-    const hasMore = rows.length > limit;
-    const sessions = rows.slice(0, limit).map(mapSession);
-    const aggregateBySession = new Map<string, AgentSessionObservationAggregate>(
-      sessions.map(session => [session.id, emptySessionObservationAggregate()]),
-    );
-
-    if (sessions.length > 0) {
-      const placeholders = sessions.map(() => '?').join(', ');
-      const aggregateRows = this.db
-        .prepare(`
-          SELECT session_id, event_type, status, late_arrival, COUNT(*) AS count
-          FROM agent_observation
-          WHERE session_id IN (${placeholders})
-          GROUP BY session_id, event_type, status, late_arrival
-        `)
-        .all(...sessions.map(session => session.id)) as Array<{
-          session_id: string;
-          event_type: string;
-          status: string;
-          late_arrival: number;
-          count: number;
-        }>;
-      for (const row of aggregateRows) {
-        const aggregate = aggregateBySession.get(row.session_id)!;
-        aggregate.total += row.count;
-        if (row.late_arrival === 1) aggregate.late += row.count;
-        aggregate.byEventType[row.event_type] =
-          (aggregate.byEventType[row.event_type] ?? 0) + row.count;
-        aggregate.byStatus[row.status] = (aggregate.byStatus[row.status] ?? 0) + row.count;
-        if (row.status === 'REDACTED') aggregate.redacted += row.count;
-        if (row.status === 'DROPPED') aggregate.dropped += row.count;
-        if (row.status === 'DEGRADED') aggregate.degraded += row.count;
-      }
-    }
-
-    return {
-      items: sessions.map(session => ({
-        session,
-        aggregate: aggregateBySession.get(session.id)!,
-      })),
-      nextCursor: hasMore && sessions.length > 0
-        ? encodeSessionCursor(sessions[sessions.length - 1]!)
-        : null,
-    };
+  ) {
+    return this.sessions.list(query);
   }
 
   getDashboardAggregate(): AgentDashboardAggregate {
-    const sessionRows = this.db
-      .prepare(`
-        SELECT status, COUNT(*) AS count
-        FROM agent_session
-        GROUP BY status
-      `)
-      .all() as Array<{ status: string; count: number }>;
-    const observationRows = this.db
-      .prepare(`
-        SELECT event_type, status, late_arrival, COUNT(*) AS count
-        FROM agent_observation
-        GROUP BY event_type, status, late_arrival
-      `)
-      .all() as Array<{
-        event_type: string;
-        status: string;
-        late_arrival: number;
-        count: number;
-      }>;
-    const aggregate: AgentDashboardAggregate = {
-      sessionsTotal: 0,
-      sessionsByStatus: {},
-      observationsTotal: 0,
-      observationsByStatus: {},
-      observationsByEventType: {},
-      redactedTotal: 0,
-      droppedTotal: 0,
-      degradedTotal: 0,
-      lateTotal: 0,
-    };
-
-    for (const row of sessionRows) {
-      aggregate.sessionsTotal += row.count;
-      aggregate.sessionsByStatus[row.status] = row.count;
-    }
-    for (const row of observationRows) {
-      aggregate.observationsTotal += row.count;
-      aggregate.observationsByStatus[row.status] =
-        (aggregate.observationsByStatus[row.status] ?? 0) + row.count;
-      aggregate.observationsByEventType[row.event_type] =
-        (aggregate.observationsByEventType[row.event_type] ?? 0) + row.count;
-      if (row.status === 'REDACTED') aggregate.redactedTotal += row.count;
-      if (row.status === 'DROPPED') aggregate.droppedTotal += row.count;
-      if (row.status === 'DEGRADED') aggregate.degradedTotal += row.count;
-      if (row.late_arrival === 1) aggregate.lateTotal += row.count;
-    }
-
-    return aggregate;
+    return this.sessions.getDashboardAggregate();
   }
 
   updateSession(
     id: string,
-    patch: Partial<Pick<
-      AgentSession,
-      'status' | 'endedAt' | 'lastEventAt' | 'maxSequenceNo' | 'degradedReason'
-    >>,
+    patch: Parameters<AgentIntegrationSessionStore['update']>[1],
     now: string,
   ): AgentSession {
-    const current = this.getSession(id);
-    if (!current) {
-      throw new Error(`Agent session not found: ${id}`);
-    }
-    this.db.prepare(`
-      UPDATE agent_session
-      SET status = ?,
-          ended_at = ?,
-          last_event_at = ?,
-          max_sequence_no = ?,
-          degraded_reason = ?,
-          updated_at = ?
-      WHERE id = ?
-    `).run(
-      patch.status ?? current.status,
-      patch.endedAt === undefined ? current.endedAt : patch.endedAt,
-      patch.lastEventAt ?? current.lastEventAt,
-      patch.maxSequenceNo ?? current.maxSequenceNo,
-      patch.degradedReason === undefined ? current.degradedReason : patch.degradedReason,
-      now,
-      id,
-    );
-    return this.getSession(id)!;
+    return this.sessions.update(id, patch, now);
   }
 
   findObservationByIdempotencyKey(
     adapterName: string,
     eventId: string,
   ): AgentObservation | null {
-    const row = this.db
-      .prepare('SELECT * FROM agent_observation WHERE adapter_name = ? AND event_id = ?')
-      .get(adapterName, eventId) as ObservationRow | undefined;
-    return row ? mapObservation(row) : null;
+    return this.observations.findByIdempotencyKey(adapterName, eventId);
   }
 
   createObservation(input: CreateObservationInput): AgentObservation {
-    this.db.prepare(`
-      INSERT INTO agent_observation (
-        id, adapter_name, event_id, session_id, event_type, sequence_no,
-        tool_name, outcome, payload_json, payload_sha256,
-        redaction_metadata_json, status, drop_reason, late_arrival,
-        occurred_at, received_at, expires_at
-      ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-    `).run(
-      input.id,
-      input.adapterName,
-      input.eventId,
-      input.sessionId,
-      input.eventType,
-      input.sequenceNo,
-      input.toolName ?? null,
-      input.outcome ?? null,
-      input.payloadJson,
-      input.payloadSha256,
-      input.redactionMetadataJson,
-      input.captureStatus,
-      input.dropReason ?? null,
-      input.lateArrival ? 1 : 0,
-      input.occurredAt,
-      input.receivedAt,
-      input.expiresAt,
-    );
-    return this.getObservation(input.id)!;
+    return this.observations.create(input);
   }
 
   getObservation(id: string): AgentObservation | null {
-    const row = this.db.prepare('SELECT * FROM agent_observation WHERE id = ?').get(id) as
-      | ObservationRow
-      | undefined;
-    return row ? mapObservation(row) : null;
+    return this.observations.get(id);
   }
 
   countObservations(sessionId: string): number {
-    const row = this.db
-      .prepare('SELECT COUNT(*) AS count FROM agent_observation WHERE session_id = ?')
-      .get(sessionId) as { count: number };
-    return row.count;
+    return this.observations.count(sessionId);
   }
 
   listObservations(
     sessionId: string,
-    query: {
-      cursor?: string;
-      limit?: number;
-      eventType?: string;
-      status?: string;
-      from?: string;
-      to?: string;
-    } = {},
+    query: Parameters<AgentIntegrationObservationStore['list']>[1] = {},
   ): ObservationPage {
-    const limit = Math.min(Math.max(query.limit ?? 50, 1), 100);
-    const conditions = ['session_id = ?'];
-    const parameters: unknown[] = [sessionId];
-    if (query.eventType) {
-      conditions.push('event_type = ?');
-      parameters.push(query.eventType);
-    }
-    if (query.status) {
-      conditions.push('status = ?');
-      parameters.push(query.status);
-    }
-    if (query.from) {
-      conditions.push('occurred_at >= ?');
-      parameters.push(query.from);
-    }
-    if (query.to) {
-      conditions.push('occurred_at <= ?');
-      parameters.push(query.to);
-    }
-    if (query.cursor) {
-      const [sequenceNo, occurredAt, receivedAt, id] = decodeCursor(query.cursor);
-      conditions.push(`
-        (sequence_no > ?)
-        OR (sequence_no = ? AND occurred_at > ?)
-        OR (sequence_no = ? AND occurred_at = ? AND received_at > ?)
-        OR (sequence_no = ? AND occurred_at = ? AND received_at = ? AND id > ?)
-      `);
-      parameters.push(
-        sequenceNo,
-        sequenceNo, occurredAt,
-        sequenceNo, occurredAt, receivedAt,
-        sequenceNo, occurredAt, receivedAt, id,
-      );
-    }
-
-    const where = conditions.map(condition => `(${condition})`).join(' AND ');
-    const rows = this.db
-      .prepare(`
-        SELECT * FROM agent_observation
-        WHERE ${where}
-        ORDER BY sequence_no, occurred_at, received_at, id
-        LIMIT ?
-      `)
-      .all(...parameters, limit + 1) as ObservationRow[];
-    const hasMore = rows.length > limit;
-    const items = rows.slice(0, limit).map(mapObservation);
-    const aggregateRows = this.db
-      .prepare(`
-        SELECT event_type, status, late_arrival, COUNT(*) AS count
-        FROM agent_observation
-        WHERE session_id = ?
-        GROUP BY event_type, status, late_arrival
-      `)
-      .all(sessionId) as Array<{
-        event_type: string;
-        status: string;
-        late_arrival: number;
-        count: number;
-      }>;
-    const aggregate = {
-      total: 0,
-      late: 0,
-      byEventType: {} as Record<string, number>,
-      byStatus: {} as Record<string, number>,
-    };
-    for (const row of aggregateRows) {
-      aggregate.total += row.count;
-      if (row.late_arrival === 1) aggregate.late += row.count;
-      aggregate.byEventType[row.event_type] =
-        (aggregate.byEventType[row.event_type] ?? 0) + row.count;
-      aggregate.byStatus[row.status] = (aggregate.byStatus[row.status] ?? 0) + row.count;
-    }
-
-    return {
-      items,
-      nextCursor: hasMore && items.length > 0 ? encodeCursor(items[items.length - 1]!) : null,
-      aggregate,
-    };
+    return this.observations.list(sessionId, query);
   }
 
   listAllObservations(sessionId: string): AgentObservation[] {
-    return (
-      this.db.prepare(`
-        SELECT * FROM agent_observation
-        WHERE session_id = ?
-        ORDER BY sequence_no, occurred_at, received_at, id
-      `).all(sessionId) as ObservationRow[]
-    ).map(mapObservation);
+    return this.observations.listAll(sessionId);
   }
 
   findPromotionCandidateByFingerprint(
     fingerprint: string,
   ): AgentMemoryPromotionCandidate | null {
-    const row = this.db.prepare(`
-      SELECT * FROM agent_memory_promotion_candidate WHERE fingerprint = ?
-    `).get(fingerprint) as PromotionCandidateRow | undefined;
-    return row ? mapPromotionCandidate(row) : null;
+    return this.promotions.findByFingerprint(fingerprint);
   }
 
   createPromotionCandidate(
     input: CreateAgentMemoryPromotionCandidateInput,
   ): AgentMemoryPromotionCandidate {
-    this.db.prepare(`
-      INSERT INTO agent_memory_promotion_candidate (
-        id, fingerprint, session_id, summary_memory_id, target_type, category,
-        content, confidence, evidence_observation_ids_json, merge_target_memory_id,
-        status, created_at, updated_at
-      ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, 'pending', ?, ?)
-    `).run(
-      input.id,
-      input.fingerprint,
-      input.sessionId,
-      input.summaryMemoryId,
-      input.targetType,
-      input.category,
-      input.content,
-      input.confidence,
-      JSON.stringify(input.evidenceObservationIds),
-      input.mergeTargetMemoryId,
-      input.createdAt,
-      input.createdAt,
-    );
-    return this.findPromotionCandidateByFingerprint(input.fingerprint)!;
+    return this.promotions.create(input);
   }
 
-  listPromotionCandidates(query: {
-    sessionId?: string;
-    status?: AgentMemoryPromotionCandidate['status'];
-  } = {}): AgentMemoryPromotionCandidate[] {
-    const conditions: string[] = [];
-    const parameters: string[] = [];
-    if (query.sessionId) {
-      conditions.push('session_id = ?');
-      parameters.push(query.sessionId);
-    }
-    if (query.status) {
-      conditions.push('status = ?');
-      parameters.push(query.status);
-    }
-    const where = conditions.length > 0 ? `WHERE ${conditions.join(' AND ')}` : '';
-    return (
-      this.db.prepare(`
-        SELECT * FROM agent_memory_promotion_candidate
-        ${where}
-        ORDER BY confidence DESC, created_at, id
-      `).all(...parameters) as PromotionCandidateRow[]
-    ).map(mapPromotionCandidate);
+  listPromotionCandidates(
+    query: Parameters<AgentIntegrationPromotionStore['list']>[0] = {},
+  ): AgentMemoryPromotionCandidate[] {
+    return this.promotions.list(query);
   }
 
-  findScopedMemoryByContent(input: {
-    targetType: AgentMemoryPromotionCandidate['targetType'];
-    content: string;
-    ownerId: string | null;
-    projectId: string | null;
-    processId: string | null;
-  }): string | null {
-    const rows = this.db.prepare(`
-      SELECT id, content FROM memory_item
-      WHERE type = ?
-        AND owner_id IS ?
-        AND project_id IS ?
-        AND process_id IS ?
-    `).all(
-      input.targetType,
-      input.ownerId,
-      input.projectId,
-      input.processId,
-    ) as Array<{ id: string; content: string }>;
-    const normalized = normalizePromotionContent(input.content);
-    return rows.find(row => normalizePromotionContent(row.content) === normalized)?.id ?? null;
+  findScopedMemoryByContent(
+    input: Parameters<AgentIntegrationPromotionStore['findScopedMemoryByContent']>[0],
+  ): string | null {
+    return this.promotions.findScopedMemoryByContent(input);
   }
 
   approvePromotionCandidate(
@@ -682,70 +142,7 @@ export class SqliteAgentIntegrationRepository implements AgentIntegrationReposit
     memoryId: string,
     now: string,
   ): AgentMemoryPromotionCandidate {
-    return this.runInTransaction(() => {
-      const row = this.db.prepare(`
-        SELECT * FROM agent_memory_promotion_candidate WHERE id = ?
-      `).get(candidateId) as PromotionCandidateRow | undefined;
-      if (!row) throw new Error(`Agent memory promotion candidate not found: ${candidateId}`);
-      const candidate = mapPromotionCandidate(row);
-      if (candidate.status !== 'pending') return candidate;
-      const session = this.getSession(candidate.sessionId);
-      if (!session) throw new Error(`Agent session not found: ${candidate.sessionId}`);
-
-      const targetExists = Boolean(
-        this.db.prepare('SELECT 1 FROM memory_item WHERE id = ?').get(memoryId),
-      );
-      if (!targetExists) {
-        this.db.prepare(`
-          INSERT INTO memory_item (
-            id, type, content, importance, privacy_scope, tags, source, origin_source,
-            owner_id, process_id, session_id, project_id, source_session_id,
-            confidence, created_at
-          ) VALUES (?, ?, ?, ?, 'private', ?, 'agent_memory_promotion', ?, ?, ?, ?, ?, ?, ?, ?)
-        `).run(
-          memoryId,
-          candidate.targetType,
-          candidate.content,
-          candidate.confidence,
-          JSON.stringify(['agent-memory', 'promotion', candidate.category]),
-          JSON.stringify({
-            tool: 'agent-memory-promotion',
-            caller: 'review',
-            timestamp: now,
-          }),
-          session.ownerId,
-          session.processId,
-          session.id,
-          session.projectId,
-          session.id,
-          candidate.confidence,
-          now,
-        );
-      }
-      this.db.prepare(`
-        INSERT OR IGNORE INTO memory_link (
-          source_id, target_id, relation_type, created_at
-        ) VALUES (?, ?, 'derived_from', ?)
-      `).run(memoryId, candidate.summaryMemoryId, now);
-      const insertProvenance = this.db.prepare(`
-        INSERT OR IGNORE INTO memory_provenance (
-          id, memory_id, session_id, observation_id, derivation_type, created_at
-        ) VALUES (?, ?, ?, ?, 'promotion', ?)
-      `);
-      for (const observationId of candidate.evidenceObservationIds) {
-        insertProvenance.run(randomUUID(), memoryId, candidate.sessionId, observationId, now);
-      }
-      this.db.prepare(`
-        UPDATE agent_memory_promotion_candidate
-        SET status = 'approved', memory_id = ?, reviewed_at = ?, updated_at = ?
-        WHERE id = ? AND status = 'pending'
-      `).run(memoryId, now, now, candidateId);
-      return mapPromotionCandidate(
-        this.db.prepare(`
-          SELECT * FROM agent_memory_promotion_candidate WHERE id = ?
-        `).get(candidateId) as PromotionCandidateRow,
-      );
-    });
+    return this.promotions.approve(candidateId, memoryId, now);
   }
 
   rejectPromotionCandidate(
@@ -753,174 +150,41 @@ export class SqliteAgentIntegrationRepository implements AgentIntegrationReposit
     reason: string,
     now: string,
   ): AgentMemoryPromotionCandidate {
-    return this.runInTransaction(() => {
-      const row = this.db.prepare(`
-        SELECT * FROM agent_memory_promotion_candidate WHERE id = ?
-      `).get(candidateId) as PromotionCandidateRow | undefined;
-      if (!row) throw new Error(`Agent memory promotion candidate not found: ${candidateId}`);
-      const candidate = mapPromotionCandidate(row);
-      if (candidate.status !== 'pending') return candidate;
-      this.db.prepare(`
-        UPDATE agent_memory_promotion_candidate
-        SET status = 'rejected', rejection_reason = ?, reviewed_at = ?, updated_at = ?
-        WHERE id = ? AND status = 'pending'
-      `).run(reason, now, now, candidateId);
-      return mapPromotionCandidate(
-        this.db.prepare(`
-          SELECT * FROM agent_memory_promotion_candidate WHERE id = ?
-        `).get(candidateId) as PromotionCandidateRow,
-      );
-    });
+    return this.promotions.reject(candidateId, reason, now);
   }
 
-  persistSessionSummary(input: {
-    memoryId: string;
-    session: AgentSession;
-    content: string;
-    observationIds: string[];
-    createdAt: string;
-  }): { memoryId: string; created: boolean } {
-    return this.runInTransaction(() => {
-      const current = this.getSession(input.session.id);
-      if (!current) throw new Error(`Agent session not found: ${input.session.id}`);
-      if (current.summaryMemoryId) {
-        return { memoryId: current.summaryMemoryId, created: false };
-      }
-      if (!['COMPLETED', 'DEGRADED', 'ABANDONED'].includes(current.status)) {
-        throw new Error(`Agent session is not terminal: ${input.session.id}`);
-      }
-
-      this.db.prepare(`
-        INSERT INTO memory_item (
-          id, type, content, importance, privacy_scope, tags, source, origin_source,
-          owner_id, process_id, session_id, project_id, source_session_id, created_at
-        ) VALUES (?, 'episodic', ?, 0.7, 'private', ?, 'agent_session_summary', ?, ?, ?, ?, ?, ?, ?)
-      `).run(
-        input.memoryId,
-        input.content,
-        JSON.stringify(['agent-session', 'summary', current.adapterName]),
-        JSON.stringify({
-          tool: 'agent-session-summary',
-          caller: 'agent-integration',
-          timestamp: input.createdAt,
-        }),
-        current.ownerId,
-        current.processId,
-        current.id,
-        current.projectId,
-        current.id,
-        input.createdAt,
-      );
-
-      const insertProvenance = this.db.prepare(`
-        INSERT INTO memory_provenance (
-          id, memory_id, session_id, observation_id, derivation_type, created_at
-        ) VALUES (?, ?, ?, ?, 'summary', ?)
-      `);
-      for (const observationId of input.observationIds) {
-        insertProvenance.run(
-          randomUUID(),
-          input.memoryId,
-          current.id,
-          observationId,
-          input.createdAt,
-        );
-      }
-
-      this.db.prepare(`
-        UPDATE agent_session
-        SET summary_memory_id = ?, updated_at = ?
-        WHERE id = ? AND summary_memory_id IS NULL
-      `).run(input.memoryId, input.createdAt, current.id);
-      return { memoryId: input.memoryId, created: true };
-    });
+  persistSessionSummary(
+    input: Parameters<AgentIntegrationPromotionStore['persistSessionSummary']>[0],
+  ): { memoryId: string; created: boolean } {
+    return this.promotions.persistSessionSummary(input);
   }
 
   markExpiredSessionsAbandoned(cutoff: string, now: string): string[] {
-    const rows = this.db.prepare(`
-      UPDATE agent_session
-      SET status = 'ABANDONED', ended_at = last_event_at, updated_at = ?
-      WHERE (
-          status IN ('ACTIVE', 'COMPACTING')
-          OR (status = 'DEGRADED' AND ended_at IS NULL)
-        )
-        AND last_event_at < ?
-      RETURNING id
-    `).all(now, cutoff) as Array<{ id: string }>;
-    return rows.map(row => row.id);
+    return this.sessions.markExpiredAbandoned(cutoff, now);
   }
 
   clearExpiredObservationPayloads(cutoff: string): number {
-    return this.db.prepare(`
-      UPDATE agent_observation
-      SET payload_json = NULL
-      WHERE expires_at IS NOT NULL
-        AND expires_at < ?
-        AND payload_json IS NOT NULL
-    `).run(cutoff).changes;
+    return this.observations.clearExpiredPayloads(cutoff);
   }
 
-  createProvenance(input: {
-    id: string;
-    memoryId: string;
-    sessionId?: string | null;
-    observationId?: string | null;
-    derivationType: string;
-    createdAt: string;
-  }): MemoryProvenance {
-    this.db.prepare(`
-      INSERT INTO memory_provenance (
-        id, memory_id, session_id, observation_id, derivation_type, created_at
-      ) VALUES (?, ?, ?, ?, ?, ?)
-    `).run(
-      input.id,
-      input.memoryId,
-      input.sessionId ?? null,
-      input.observationId ?? null,
-      input.derivationType,
-      input.createdAt,
-    );
-    const row = this.db.prepare('SELECT * FROM memory_provenance WHERE id = ?').get(input.id) as
-      ProvenanceRow;
-    return mapProvenance(row);
+  createProvenance(
+    input: Parameters<AgentIntegrationProvenanceStore['create']>[0],
+  ): MemoryProvenance {
+    return this.provenance.create(input);
   }
 
-  listProvenance(query: {
-    memoryId?: string;
-    observationId?: string;
-  }): MemoryProvenance[] {
-    const conditions: string[] = [];
-    const parameters: string[] = [];
-    if (query.memoryId) {
-      conditions.push('memory_id = ?');
-      parameters.push(query.memoryId);
-    }
-    if (query.observationId) {
-      conditions.push('observation_id = ?');
-      parameters.push(query.observationId);
-    }
-    if (conditions.length === 0) return [];
-    return (
-      this.db
-        .prepare(`SELECT * FROM memory_provenance WHERE ${conditions.join(' OR ')}`)
-        .all(...parameters) as ProvenanceRow[]
-    ).map(mapProvenance);
+  listProvenance(
+    query: Parameters<AgentIntegrationProvenanceStore['list']>[0],
+  ): MemoryProvenance[] {
+    return this.provenance.list(query);
   }
 
   markProvenanceSourceDeleted(sessionId: string): number {
-    return this.db.prepare(`
-      UPDATE memory_provenance
-      SET source_deleted = 1
-      WHERE session_id = ?
-         OR observation_id IN (
-           SELECT id FROM agent_observation WHERE session_id = ?
-         )
-    `).run(sessionId, sessionId).changes;
+    return this.provenance.markSourceDeleted(sessionId);
   }
 
   deleteSession(sessionId: string): boolean {
-    this.markProvenanceSourceDeleted(sessionId);
-    return this.db.prepare('DELETE FROM agent_session WHERE id = ?').run(sessionId).changes > 0;
+    return this.sessions.delete(sessionId, id => this.provenance.markSourceDeleted(id));
   }
 
   exportSession(sessionId: string): {
@@ -928,25 +192,19 @@ export class SqliteAgentIntegrationRepository implements AgentIntegrationReposit
     observations: AgentObservation[];
     provenance: MemoryProvenance[];
   } | null {
-    const session = this.getSession(sessionId);
+    const session = this.sessions.get(sessionId);
     if (!session) return null;
     const observations: AgentObservation[] = [];
     let cursor: string | undefined;
     do {
-      const page = this.listObservations(sessionId, { limit: 100, cursor });
+      const page = this.observations.list(sessionId, { limit: 100, cursor });
       observations.push(...page.items);
       cursor = page.nextCursor ?? undefined;
     } while (cursor);
-    const provenance = (
-      this.db.prepare(`
-        SELECT * FROM memory_provenance
-        WHERE session_id = ?
-           OR observation_id IN (
-             SELECT id FROM agent_observation WHERE session_id = ?
-           )
-      `).all(sessionId, sessionId) as
-        ProvenanceRow[]
-    ).map(mapProvenance);
-    return { session, observations, provenance };
+    return {
+      session,
+      observations,
+      provenance: this.provenance.listForSession(sessionId),
+    };
   }
 }
